messages

+++ app/src/main/java/com/example/getgoing/Event.kt
@@ -1,3 +1,3 @@
 package com.example.getgoing
 
-data class Event(var name: String, var totalAmount: Int)
+data class Event(var name: String, var totalAmount: Double)
